moved '/util/json/json.go' -> '/util/json.go', fixed wrong package decl
